Support fancy markdown slides, and image credit statements
The code for the two converted slides looks pretty complicated, and one
could argue that the html version is just simpler. However, that is only
because the result tries to be feature equivalent with the previous HTML
version.  When cranking out slides for lectures the desire is often to
tweak as little as possible, and I quickly stick to the default
components: images, lists, enumerations, code snippets and math. When
sticking to those bits that are well defined in markdown, the time it
takes to compose a slide in markdown (just the typing) is massively
reduced, and the focus is on the content, not the markup and layout.

I added default formatting of image credit statements (via a custom
tag). Those are important, but they are not the primary content of a
slide, hence I made them less visible and off to the side. The dedicated
tag marks those statements up semantically, and different from a general
slide note (which are useful for citations and other references related
to the primary message of a slide).